+++ src/qgis_plugin_dev_tools/translations/__init__.py
@@ -5,6 +5,7 @@ from pathlib import Path
 
 from qgis_plugin_dev_tools.translations.update_translations import (
     get_unfinished_translations_count,
+    run_command,
     update_ts_file,
 )
 
@@ -50,3 +51,19 @@ def update_translation_files(
                 shutil.copy(backup_ts_file, ts_file)
             else:
                 LOGGER.info("Updated translations in %s", ts_file)
+
+
+def compile_translations(
+    language_codes: list[str],
+    source_path: Path,
+) -> None:
+    for language_code in language_codes:
+        ts_file = source_path / f"{language_code}.ts"
+        qm_file = source_path / f"{language_code}.qm"
+        if ts_file.exists():
+            LOGGER.debug("compiling %s into %s...", ts_file, qm_file)
+            run_command(["lrelease", str(ts_file), "-qm", str(qm_file)])
+        else:
+            LOGGER.warning(
+                "No translation file found for %s in %s", language_code, source_path
+            )